Add a cast to avoid warning on 64-bits.

+++ Var1/Glasses/TriMesh.cxx
@@ -213,7 +213,7 @@ void TriMesh::ExportGTSurf(GTSurf* gts)
       ta += 3;
     }
   }
-  printf("Counting edges => %d, map-size = %d\n", edge_cnt, edge_map.size());
+  printf("Counting edges => %d, map-size = %d\n", edge_cnt, (int)edge_map.size());
 
   // Construct surface
